pyproject.toml
[project]
name = "cenacellm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.3.0",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.7",
    "mistralai>=1.5.0",
    "numpy>=2.2.3",
    "ollama>=0.4.7",
    "openai>=1.65.3",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "pypdf>=5.3.1",
    "semantic-text-splitter>=0.24.1",
]

[project.scripts]
cenacellm = "cenacellm:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.uv.workspace]
members = ["proyectoCenace"]
